requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
exceptiongroup==1.2.0
fastapi==0.110.1
gotrue==2.4.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
packaging==24.0
postgrest==0.16.3
pydantic==2.7.0
pydantic_core==2.18.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.3
supafunc==0.4.5
typing_extensions==4.11.0
uvicorn==0.29.0
websockets==12.0
